edmonds-karp creates empty paths

+++ includes/info.h
@@ -6,7 +6,7 @@
 /*   By: carlnysten <marvin@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/27 21:17:07 by carlnysten        #+#    #+#             */
-/*   Updated: 2022/07/07 11:08:13 by carlnysten       ###   ########.fr       */
+/*   Updated: 2022/07/12 11:27:44 by cnysten          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ typedef struct s_info
 	char	*source;
 	char	*sink;
 	long	n_ants;
-	long	max_flow;
+	int		max_flow;
 }	t_info;
 
 #endif
